% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokedex_df.R
\docType{data}
\name{pokedex_df}
\alias{pokedex_df}
\title{Size measurements for adult foraging penguins near Palmer Station, Antarctica}
\format{
A tibble with 386 rows and 13 variables:
}
\usage{
pokedex_df
}
\description{
Includes measurements for penguin species, island in Palmer Archipelago,
size (flipper length, body mass, bill dimensions), and sex.
}
\keyword{datasets}
