% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokegen.R
\name{pokegen}
\alias{pokegen}
\title{Generate a team of pokemons}
\usage{
pokegen(n = 6)
}
\arguments{
\item{n}{Integer. Number of pokemons to generate. Can be from 1 pokemon to 6 pokemons}
}
\value{
Vector of names of the \code{n} pokemons generated
}
\description{
\code{pokegen} creates a vector of \code{n} random pokemons
}
\examples{
pokegen(3)
}
