% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokestats.R
\name{pokestats}
\alias{pokestats}
\title{Title}
\usage{
pokestats(
  myteam = c("golem", "staryu", "exeggutor", "wynaut", "wailmer", "xatu"),
  title = "My Pokemon Team"
)
}
\arguments{
\item{title}{}
}
\description{
Title
}
