% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokecard.R
\name{pokecard}
\alias{pokecard}
\title{Title}
\usage{
pokecard(
  myteam = c("pikachu", "staryu", "exeggutor", "wynaut", "wailmer", "xatu"),
  color = "#c60031",
  title = "My Pokemon Team"
)
}
\arguments{
\item{title}{}
}
\description{
Title
}
