% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokecard.R
\name{pokecard}
\alias{pokecard}
\title{Create a flashcard of the pokemon team}
\usage{
pokecard(myteam = pokegen(), title = "My Pokemon Team", color = "#c60031")
}
\arguments{
\item{myteam}{A vector with a length of up 6 to pokemons. It can be valid numeric ids or valid pokemon names (lowercase). By default, it creates a flashcard of a team of 6 random pokemons.}

\item{title}{A string. It determines the title at the top of the flashcard. By default, the title is 'My Pokemon Team'}

\item{color}{A string, which should be a valid color. It determines the color of the border of the flashcard.}
}
\value{
An external pointer of class 'magick-image'. It shows the flashcard along with a tibble describing the features of the image.
}
\description{
\code{pokecard} create an image of the pokemon team, appending their sprites
}
\examples{
pokecard(c("pikachu","bulbasaur", "squirtle", "charmander", "pidgey", "psyduck"))
pokecard(c("charmander","charmeleon", "charizard"),  "Charmander's evolutions", "Orange")
}
