% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokedex.R
\name{pokedex}
\alias{pokedex}
\title{Obtain the information and stats of the pokemons in a data frame}
\usage{
pokedex(ids = seq(1, 386))
}
\arguments{
\item{ids}{Vector of integers (from 1 to 386) or valid pokemon names (lowercase). By default, it extracts the information and stats of the pokemons of the first 3 generations, which corresponds to the numeric ids 1 to 386}
}
\value{
Data frame with the information and stats of the pokemons
}
\description{
\code{pokedex} creates a dataframe with information and stats of the corresponding pokemons
}
\examples{
pokedex(c("bulbasaur", "charmander", "squirtle"))
pokedex(c(1,4,7))
}
