% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokestats.R
\name{pokestats}
\alias{pokestats}
\title{Barplot of the main stats of the pokemons}
\usage{
pokestats(myteam = pokegen(), title = "My Pokemon Team")
}
\arguments{
\item{myteam}{A vector with a length of up 6 to pokemons. It can be valid numeric ids or valid pokemon names (lowercase). By default, it creates a barplot of a team of 6 random pokemons.}

\item{title}{A string. It determines the title at the top of the plot By default, the title is 'My Pokemon Team'}
}
\value{
ggplot2 object, which contains the barplot of the main stats for each pokemon
}
\description{
\code{pokestats} creates a barplot of the main stats of the pokemons (HP, Attack, Defense, Sp. Attack, Sp. Defense and Speed)
}
\examples{
pokestats(myteam = c("pikachu","bulbasaur", "squirtle", "charmander", "pidgey", "psyduck"))
pokestats(myteam = c("charmander","charmeleon", "charizard"))
}
